third_party/blink/web_tests/external/wpt/css/css-variables/variable-definition-expected.txt

This is a testharness.js-based test.
Found 73 tests; 44 PASS, 29 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS no variable
PASS variable
PASS single char variable
PASS single char '-' variable
PASS no char variable
FAIL white space value (single space) assert_equals: Expected Value should match actual value expected "" but got " "
FAIL white space value (double space) assert_equals: Expected Value should match actual value expected "" but got " "
PASS overwrite
PASS can overwrite with no value
FAIL can overwrite with space value assert_equals: Expected Value should match actual value expected "" but got " "
PASS case sensetivity
PASS case sensetivity2
PASS parsing three dashes at start of variable
PASS parsing multiple dashes with one dash at start of variable
FAIL  leading white space (single space) assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (single space)
FAIL  trailing white space (single space) assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL  leading white space (double space) 2 assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (double space) 2
FAIL  trailing white space (double space) 2 assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL !important assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS !important 2
FAIL !important (with space) assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS no variable (Computed Style)
PASS variable (Computed Style)
PASS single char variable (Computed Style)
PASS single char '-' variable (Computed Style)
PASS no char variable (Computed Style)
FAIL white space value (single space) (Computed Style) assert_equals: Expected Value should match actual value expected "" but got " "
FAIL white space value (double space) (Computed Style) assert_equals: Expected Value should match actual value expected "" but got " "
PASS overwrite (Computed Style)
PASS can overwrite with no value (Computed Style)
FAIL can overwrite with space value (Computed Style) assert_equals: Expected Value should match actual value expected "" but got " "
PASS case sensetivity (Computed Style)
PASS case sensetivity2 (Computed Style)
PASS parsing three dashes at start of variable (Computed Style)
PASS parsing multiple dashes with one dash at start of variable (Computed Style)
FAIL  leading white space (single space) (Computed Style) assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (single space) (Computed Style)
FAIL  trailing white space (single space) (Computed Style) assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL  leading white space (double space) 2 (Computed Style) assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (double space) 2 (Computed Style)
FAIL  trailing white space (double space) 2 (Computed Style) assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL !important (Computed Style) assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS !important 2 (Computed Style)
FAIL !important (with space) (Computed Style) assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS no variable (Cascading)
PASS variable (Cascading)
PASS single char variable (Cascading)
PASS single char '-' variable (Cascading)
PASS no char variable (Cascading)
FAIL white space value (single space) (Cascading) assert_equals: Expected Value should match actual value expected "" but got " "
FAIL white space value (double space) (Cascading) assert_equals: Expected Value should match actual value expected "" but got " "
PASS overwrite (Cascading)
PASS can overwrite with no value (Cascading)
FAIL can overwrite with space value (Cascading) assert_equals: Expected Value should match actual value expected "" but got " "
PASS case sensetivity (Cascading)
PASS case sensetivity2 (Cascading)
PASS parsing three dashes at start of variable (Cascading)
PASS parsing multiple dashes with one dash at start of variable (Cascading)
FAIL  leading white space (single space) (Cascading) assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (single space) (Cascading)
FAIL  trailing white space (single space) (Cascading) assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL  leading white space (double space) 2 (Cascading) assert_equals: Expected Value should match actual value expected "value" but got " value"
PASS  middle white space (double space) 2 (Cascading)
FAIL  trailing white space (double space) 2 (Cascading) assert_equals: Expected Value should match actual value expected "value" but got "value "
FAIL !important (Cascading) assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS !important 2 (Cascading)
FAIL !important (with space) (Cascading) assert_equals: Expected Value should match actual value expected "value1" but got "value1 "
PASS basic CSSOM.setProperty
FAIL CSSOM.setProperty with space 1 assert_equals: Expected Value should match actual value expected "" but got "green"
FAIL CSSOM.setProperty with space 2 assert_equals: Expected Value should match actual value expected "" but got "green"
PASS CSSOM.setProperty with space 3
Harness: the test ran to completion.
